nn update

+++ cpp/chase/mtwister.c
@@ -72,5 +72,5 @@ unsigned long genRandLong(MTRand* rand) {
  * Generates a pseudo-randomly generated double in the range [0..1].
  */
 double genRand(MTRand* rand) {
-  return((double)genRandLong(rand) / (unsigned long)0xffffffff);
+  return (double)genRandLong(rand) / (double)0xffffffff;
 }
